requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
watchdog>=3.0.0

# Phase 1: Database + Backend API (FREE)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
# Pin bcrypt stack for compatibility on Windows / Python 3.13
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0